Moved run generator, added run generartor initialization
I moved the Run Generator view onto it's own route and into it's own file. I created a dummy index file with a hyperlink to the run generator. I also modified the Run Generator so it starts with a generated run already instead of having to click "generate" to obtain the run.